Move ProcessInterface into ForkManager namespace.
